implementing dates and validating in service layer

+++ src/main/java/br/com/rentacarserver/rentacarserver/services/BrandService.java
@@ -6,12 +6,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
+import java.time.Instant;
 import java.util.List;
 
 @Service
 public class BrandService {
 
     private BrandRepository brandRepository;
+    protected Instant dateNow = Instant.now();
 
     public BrandService( BrandRepository brandRepository){
         this.brandRepository = brandRepository;
@@ -26,23 +28,41 @@ public class BrandService {
     }
 
     public void insertBrand(@RequestBody final BrandEntity brandEntity){
-        try{
-            if(brandEntity.getName() != null && (
-                    !brandEntity.getName().isEmpty())){
-                this.brandRepository.save(brandEntity);
+            if(brandEntity.getName() != null &&
+                    (!brandEntity.getName().isEmpty())){
+                if(brandEntity.getId() == null ||
+                        (!this.brandRepository.existsById(brandEntity.getId()))){
+                    brandEntity.setCreatedAt(dateNow.getEpochSecond());
+                    this.brandRepository.save(brandEntity);
+                }else {
+                    throw new IllegalArgumentException("Invalid ID");
+                }
+            }else{
+                throw new IllegalArgumentException("Name as null or empty");
             }
-        }catch (Exception e){
-            e.getMessage();
-        }
     }
 
     public void updateBrand(@RequestBody final BrandEntity brandEntity){
-        this.brandRepository.save(brandEntity);
+        if(brandEntity.getId() != null &&
+                (this.brandRepository.existsById(brandEntity.getId()))) {
+            if(brandEntity.getName() != null &&
+                (!brandEntity.getName().isEmpty())) {
+                brandEntity.setCreatedAt(dateNow.getEpochSecond());
+                this.brandRepository.save(brandEntity);
+            }else{
+                throw new IllegalArgumentException("Name can't be null");
+            }
+        }else{
+            throw new IllegalArgumentException("Update need a valid id");
+        }
     }
 
     public void deleteBrandById(@PathVariable(value = "id") Long brandId){
-        if(brandId != null) {
+        if(brandId != null &&
+                (this.brandRepository.existsById(brandId))){
             this.brandRepository.deleteById(brandId);
+        }else{
+            throw new IllegalArgumentException("id can't be null");
         }
     }
